Move the schema type enumeration to cassandra/blob_storage.hpp JIRA: CXX-10436
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@85157 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/app/pubseq_gateway/pubseq_gateway.cpp
@@ -639,7 +639,7 @@ int CPubseqGatewayApp::x_PopulateSatToKeyspaceMap(void)
     try {
         string      err_msg;
         FetchSatToKeyspaceMapping(m_RootKeyspace, m_CassConnection,
-                                  m_SatNames, static_cast<int>(eBlobVer2Schema),
+                                  m_SatNames, eBlobVer2Schema,
                                   m_BioseqKeyspace, eResolverSchema,
                                   err_msg);
         if (!err_msg.empty()) {
